LPS-67573 Move internal classes to their own packages
